feat: improve city toggle
